Fixing errors

+++ denspp/offline/analog/amplifier/pre_amp.py
@@ -24,7 +24,7 @@ class SettingsAMP:
     vss:    float
     fs_ana: float
     # Amplifier characteristics
-    gain:   int
+    gain:   float
     n_filt: int
     f_filt: list
     f_type: str
@@ -42,7 +42,7 @@ class SettingsAMP:
 
 RecommendedSettingsAMP = SettingsAMP(
     vdd=0.6, vss=-0.6,
-    fs_ana=50e3, gain=40,
+    fs_ana=50e3, gain=40.,
     n_filt=1, f_filt=[0.1, 8e3], f_type="bandpass",
     offset=0e-6,
     f_chop=10e3,
